Move HomePage component into its own subdirectory and add stories
