+++ src/pages/homePage/homePage.module.css
@@ -17,6 +17,8 @@
 }
 
 .login {
+  font-family: 'Quattrocento Sans', Arial, Helvetica, sans-serif;
+  font-size: 1.1rem;
   color: #fff;
   text-decoration: none;
   margin: 0 auto;
