+++ src/pages/homePage/homePage.stories.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import HomePage from './homePage'
+
+export default { title: 'HomePage' }
+
+export const Default = () => <HomePage />
